% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_svg}
\alias{write_svg}
\title{Write a plot to fixed-size SVG}
\usage{
write_svg(plot, path, width_mm, height_mm)
}
\arguments{
\item{plot}{\code{ggplot} object to render.}

\item{path}{Destination file path.}

\item{width_mm}{Width in millimeters.}

\item{height_mm}{Height in millimeters.}
}
\value{
The written path.
}
\description{
Write a plot to fixed-size SVG
}
\examples{
coverage_df <- data.frame(
  month = as.Date(c("2024-01-01", "2024-02-01")),
  employee_count = c(12L, 15L)
)
plot <- plot_monthly_measurement_coverage(coverage_df)
path <- tempfile(fileext = ".svg")
write_svg(plot, path, width_mm = 120, height_mm = 70)
file.exists(path)
unlink(path)
}
\seealso{
\code{\link[svglite:svglite]{svglite::svglite()}}, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
